vgdb: Add multi mode
vgdb multi: Use file descriptor to set the valgrind working directory

vgdb: create a pipe in for_and_exec_valgrind

After the fork there are two processes which share
memory and file descriptors. When we close one side
of the pipe, it only closes that file descriptor in
that one process. in the parent (vgdb) we simply keep
reading till we see something bad, or the file descriptor
(on the other side) is closed.